+++ client/src/App.js
@@ -3,7 +3,6 @@ import { BrowserRouter as Router,Route, Switch} from 'react-router-dom';
 import './resources/assets/sass/app.scss';
 import './resources/assets/icomoon/icon.scss'
 import Home from './resources/assets/components/Home';
-import Other from './resources/assets/components/Other';
 
 class App extends Component {
   render() {
@@ -11,7 +10,8 @@ class App extends Component {
       <Router>
         <Switch>
           <Route exact path="/" component={Home} />
-          <Route exact path="/other" component={Other} />
+          <Route exact path="/propose/:id" component={Home} />
+          {/* <Route exact path="/user/:id" component={User} /> */}
         </Switch>
       </Router>
     );
